Move interfaces to shared folder
