+++ src/shared/interfaces/ValidationInterface.tsx
